+++ plugins/@grouparoo/demo/src/events.ts
@@ -1,10 +1,9 @@
 import { runAction } from "./util/runAction";
-import * as uuid from "uuid";
-import { ApiKey, App, ProfilePropertyRule, Source } from "@grouparoo/core";
+import { MockSession, getApiKey } from "./util/MockSession";
+import { App, ProfilePropertyRule, Source } from "@grouparoo/core";
 import { getPurchaseCategories, getPurchases } from "./sample_data";
 import { createPropertyRule, RuleDefinition } from "./util/propertyRules";
-import { log, userCreatedAt } from "./util/shared";
-import fetch from "isomorphic-fetch";
+import { log } from "./util/shared";
 
 const PAGEVIEW_RULES: Array<RuleDefinition> = [
   {
@@ -165,274 +164,4 @@ async function findIdentifyingRuleGuid() {
     throw new Error(`Identifying rule not found: ${where.key}`);
   }
   return found.guid;
-}
-
-export async function getApiKey(): Promise<ApiKey> {
-  const where = { name: "Demo API Key" };
-
-  const found = await ApiKey.findOne({ where });
-  const params = {
-    name: where.name,
-    permissionAllRead: true,
-    permissionAllWrite: true,
-    guid: found?.guid,
-    permissions: [],
-  };
-  if (found) {
-    await runAction("apiKey:edit", params);
-  } else {
-    await runAction("apiKey:create", params);
-  }
-
-  const made = await ApiKey.findOne({ where });
-  if (!made) {
-    throw new Error(`Events apiKey not made`);
-  }
-  return made;
-}
-
-// prettier-ignore
-const funnel = [
-  { type: "pageview", data: { page: "/" }, percent: 100, user:false },
-  { type: "pageview", data: { page: "/about" }, percent: 90, user: false },
-  { type: "pageview", data: { page: "/product" }, percent: 80, user: false, category: true },
-  { type: "pageview", data: { page: "/cart" }, percent: 75, user: false, category: true },
-  { type: "pageview", data: { page: "/sign-in" }, percent: 60, user: true },
- // { type: "identify", data: { page: "/sign-in" }, percent: 100 },
-  { type: "pageview", data: { page: "/purchase" }, percent: 80, user: true, category: true, purchase: true },
-  { type: "pageview", data: { page: "/thanks" }, percent: 100, user: true, category: true, purchase: true },
-];
-
-export class MockSession {
-  id: string;
-  apiKey: ApiKey;
-  purchase: { [key: string]: any };
-  category: string;
-  categories: string[];
-  anonymousId: string;
-  currentUserId: number;
-  currentStep: number;
-  identified: boolean;
-  now: number;
-  baseUrl: string;
-  done: boolean;
-  started: boolean;
-
-  constructor(
-    id: string,
-    apiKey: ApiKey,
-    purchase: { [key: string]: any },
-    categories: string[],
-    baseUrl?: string
-  ) {
-    this.id = id;
-    this.apiKey = apiKey;
-    this.anonymousId = null;
-    this.currentUserId = null;
-    this.currentStep = 0;
-    this.identified = false;
-    this.purchase = purchase;
-    this.category = null;
-    this.categories = categories;
-    this.now = 0;
-    this.baseUrl = baseUrl;
-    this.done = false;
-    this.started = false;
-  }
-
-  async start() {
-    if (!this.started) {
-      this.started = true;
-      await this.pickCurrentUser();
-      await this.setNow();
-    }
-  }
-
-  async run() {
-    while (await this.tick()) {
-      this.now += 1000;
-      log(4, `tick: ${this.id} -> ${this.currentStep}`);
-    }
-  }
-
-  async tick(): Promise<boolean> {
-    if (this.done) {
-      return false;
-    }
-    await this.start();
-    const step = funnel[this.currentStep];
-    if (!this.shouldContinue(step)) {
-      this.done = true;
-      return false;
-    }
-
-    await this.track(step);
-    this.currentStep++;
-    return true;
-  }
-
-  shouldContinue(step) {
-    if (!step) {
-      return false;
-    }
-    if (this.purchase) {
-      return true; // going to the end!
-    }
-    if (step.user && !this.currentUserId) {
-      return false; // was fully anonymous user
-    }
-    if (step.purchase && !this.purchase) {
-      return false; // not going to happen
-    }
-
-    const randomNumber = Math.floor(Math.random() * 101);
-    if (randomNumber > step.percent) {
-      return false;
-    }
-
-    // carry on!
-    return true;
-  }
-
-  pickUserId(): number {
-    return Math.floor(Math.random() * 1000) + 1;
-  }
-
-  // find a user, (or stay anonymous -- null)
-  async pickCurrentUser() {
-    const loggedOutPercentage = 12;
-    const randomNumber = Math.floor(Math.random() * 101);
-
-    let userId: number;
-    if (this.purchase) {
-      // use this user
-      userId = this.purchase.user_id;
-    } else if (randomNumber < loggedOutPercentage) {
-      // users 1 through 1000
-      userId = this.pickUserId();
-    } else {
-      // otherwise, stay logged out
-      userId = null;
-    }
-
-    this.currentUserId = userId;
-    this.anonymousId = `?-${uuid.v4()}`;
-
-    log(
-      4,
-      `${this.id} userId: ${this.currentUserId} anonymousId: ${this.anonymousId}`
-    );
-  }
-
-  // set the time to start, there's a second between each step
-  async setNow() {
-    let userId = this.currentUserId;
-    if (!userId) {
-      // random user, pick one
-      userId = this.pickUserId();
-    }
-    let firstUser = -1;
-    let firstPurchase = -1;
-
-    for (let i = 0; i < funnel.length; i++) {
-      const step = funnel[i];
-      if (firstUser < 0 && step.user) {
-        firstUser = i;
-      }
-      if (firstPurchase < 0 && step.purchase) {
-        firstPurchase = i;
-      }
-    }
-    if (this.purchase) {
-      // use that
-      this.now = this.purchase.created_at.getTime() - firstPurchase * 1000;
-    } else {
-      // some time after created
-      const now = new Date().getTime();
-      let generatedCreateAt = await userCreatedAt(userId);
-      let creationAgo = now - generatedCreateAt.getTime();
-      creationAgo = Math.random() * creationAgo;
-      this.now = now - creationAgo;
-    }
-  }
-
-  async getCategory() {
-    if (this.category) {
-      return this.category;
-    }
-    if (this.purchase) {
-      this.category = this.purchase.category;
-    } else {
-      // otherwise pick one
-      const index = Math.floor(Math.random() * this.categories.length);
-      this.category = this.categories[index];
-    }
-
-    if (!this.category) {
-      throw new Error("no category set!");
-    }
-    return this.category;
-  }
-
-  async track(step) {
-    let occurredAt: number = this.now;
-
-    const params: any = {
-      anonymousId: this.anonymousId,
-      occurredAt: occurredAt,
-      type: step.type,
-      data: step.data || {},
-    };
-
-    if (this.identified) {
-      params.userId = this.currentUserId;
-    }
-    if (step.category) {
-      params.data.category = await this.getCategory();
-    }
-
-    await this.sendEvent(params);
-
-    if (!this.identified && step.user && this.currentUserId) {
-      await this.sendEvent({
-        anonymousId: this.anonymousId,
-        userId: this.currentUserId,
-        occurredAt: occurredAt + 1,
-        type: "identify",
-        data: params.data,
-      });
-      this.identified = true;
-    }
-  }
-
-  async sendEvent(params) {
-    if (this.baseUrl) {
-      await this.postEvent(params);
-    } else {
-      await runAction("event:create", params, { apiKey: this.apiKey });
-    }
-  }
-
-  async postEvent(params) {
-    const data = Object.assign({}, params, { apiKey: this.apiKey.apiKey });
-    data.occuredAt = new Date().getTime(); // streaming, do it now
-    log(1, this.baseUrl + " " + JSON.stringify(data));
-    await fetch(this.baseUrl, {
-      method: "post",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify(data),
-    })
-      .then((response) => {
-        if (response.ok && !response.error) {
-          return response.json();
-        } else {
-          return Promise.reject(response);
-        }
-      })
-      .catch((err) => {
-        return Promise.reject(err);
-      });
-  }
 }
